requirements.txt
django
djangorestframework
django-cors-headers
pandas
xlrd
pymysql